Feat: Add support for AGPL and BSD licensed builds
Adds build tags and platform-specific logic to handle both
AGPL-licensed and BSD-licensed versions of the library.

This introduces file separations for the two license types
and updates build logic accordingly.
