+++ file.go
@@ -238,15 +238,6 @@ func (f *GoFile) SourceInfo(fn *Function) (string, int, int) {
 	return srcFile, start, end
 }
 
-// GetGoRoot returns the Go Root path used to compile the binary.
-func (f *GoFile) GetGoRoot() (string, error) {
-	err := f.initPackages()
-	if err != nil {
-		return "", err
-	}
-	return findGoRootPath(f)
-}
-
 // SetGoVersion sets the assumed compiler version that was used. This
 // can be used to force a version if gore is not able to determine the
 // compiler version used. The version string must match one of the strings
