+++ slow_test.go
@@ -14,6 +14,7 @@
 //
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
+
 //go:build slow_test
 
 package gore
@@ -317,26 +318,6 @@ func TestSourceInfo(t *testing.T) {
 		a.NotEqual(0, start)
 		a.NotEqual(0, end)
 		a.NotEqual("", file)
-	})
-}
-
-func TestDwarfString(t *testing.T) {
-	noStrip := false
-	getMatrix(t, nil, &noStrip, "dwarfString", func(t *testing.T, exe string) {
-		r := require.New(t)
-
-		f, err := Open(exe)
-		r.NoError(err)
-		r.NotNil(f)
-		defer f.Close()
-
-		gover, ok := getBuildVersionFromDwarf(f.fh)
-		r.True(ok)
-		r.Equal(gover, runtime.Version())
-
-		goroot, ok := getGoRootFromDwarf(f.fh)
-		r.True(ok)
-		r.Equal(goroot, runtime.GOROOT())
 	})
 }
